pyproject.toml
[project]
name = "gp-sphinx-workspace"
version = "0.0.1a0"
description = "Workspace root for gp-sphinx packages"
requires-python = ">=3.10,<4.0"
authors = [
  {name = "Tony Narlock", email = "tony@git-pull.com"}
]
license = { text = "MIT" }
readme = "README.md"
dependencies = [
  "gp-sphinx==0.0.1a0",
]

[tool.uv.workspace]
members = ["packages/*"]

[tool.uv.sources]
sphinx-fonts = { workspace = true }
sphinx-gptheme = { workspace = true }
sphinx-argparse-neo = { workspace = true }
sphinx-autodoc-pytest-fixtures = { workspace = true }
sphinx-autodoc-docutils = { workspace = true }
sphinx-autodoc-sphinx = { workspace = true }
gp-sphinx = { workspace = true }

[dependency-groups]
dev = [
  "gp-sphinx",
  "sphinx-argparse-neo",
  "sphinx-autodoc-pytest-fixtures",
  "sphinx-autodoc-docutils",
  "sphinx-autodoc-sphinx",
  # Docs
  "sphinx-autobuild",
  # Testing
  "pytest",
  "pytest-rerunfailures",
  "pytest-mock",
  "pytest-watcher",
  # Coverage
  "codecov",
  "coverage",
  "pytest-cov",
  # Lint
  "ruff",
  "mypy",
  "tomli; python_version < '3.11'",
  "typing-extensions",
  "types-docutils",
  "types-Pygments",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gp_sphinx_workspace"]

[tool.mypy]
strict = true
python_version = "3.10"
mypy_path = "scripts/ci:docs/_ext"
files = [
  "packages/",
  "tests/",
  "scripts/",
]

[[tool.mypy.overrides]]
# sphinx_fonts sets private attributes on Sphinx's app object (app._font_preload_hrefs,
# _font_faces, etc.) which aren't declared in sphinx's type stubs.
# The test file also accesses these dynamically-added attrs, so it needs the same suppression.
module = ["sphinx_fonts", "tests.ext.test_sphinx_fonts"]
disable_error_code = ["attr-defined", "unused-ignore"]

[[tool.mypy.overrides]]
# sphinx_autodoc_pytest_fixtures: not yet typed strictly; ignore_errors is temporary
# until targeted disable_error_code can replace it.
module = ["sphinx_autodoc_pytest_fixtures", "sphinx_autodoc_pytest_fixtures.*", "tests.ext.pytest_fixtures.*"]
ignore_errors = true

[tool.ruff]
target-version = "py310"
extend-exclude = [
  "build_extensions.py",
  "copy_packages.py",
  "rewrite_docutils.py",
  "rewrite_sphinx.py",
]

[tool.ruff.lint]
select = [
  "E", # pycodestyle
  "F", # pyflakes
  "I", # isort
  "UP", # pyupgrade
  "A", # flake8-builtins
  "B", # flake8-bugbear
  "C4", # flake8-comprehensions
  "COM", # flake8-commas
  "EM", # flake8-errmsg
  "Q", # flake8-quotes
  "PTH", # flake8-use-pathlib
  "SIM", # flake8-simplify
  "TRY", # Trycertatops
  "PERF", # Perflint
  "RUF", # Ruff-specific rules
  "D", # pydocstyle
  "FA100",  # future annotations
]
ignore = [
  "COM812", # missing trailing comma, ruff format conflict
]
extend-safe-fixes = [
  "UP006",
  "UP007",
]
pyupgrade.keep-runtime-typing = false

[tool.ruff.lint.isort]
known-first-party = [
  "gp_sphinx",
  "sphinx_fonts",
  "sphinx_gptheme",
  "sphinx_argparse_neo",
  "sphinx_autodoc_pytest_fixtures",
  "sphinx_autodoc_docutils",
  "sphinx_autodoc_sphinx",
]
combine-as-imports = true
required-imports = [
  "from __future__ import annotations",
]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.per-file-ignores]
"*/__init__.py" = ["F401"]
"docs/_ext/package_reference.py" = ["B009", "D102", "D301", "E501", "PERF401"]
"packages/sphinx-argparse-neo/**/*.py" = ["E501", "UP", "A", "B", "COM", "EM", "TRY", "PERF", "RUF", "SIM", "FA100"]
"packages/sphinx-autodoc-pytest-fixtures/**/*.py" = ["D417", "E501", "UP", "A", "B", "COM", "EM", "TRY", "PERF", "RUF", "SIM", "FA100"]
"packages/sphinx-autodoc-docutils/**/*.py" = ["D417", "E501", "UP", "A", "B", "COM", "EM", "TRY", "PERF", "RUF", "SIM", "FA100"]
"tests/ext/*.py" = ["D", "E501", "UP", "A", "B", "COM", "EM", "TRY", "PERF", "RUF", "SIM", "FA100"]
"tests/ext/argparse_neo/*.py" = ["D", "E501", "UP", "A", "B", "COM", "EM", "TRY", "PERF", "RUF", "SIM", "FA100"]
"tests/ext/autodoc_docutils/*.py" = ["D", "E501", "UP", "A", "B", "COM", "EM", "TRY", "PERF", "RUF", "SIM", "FA100"]
"tests/ext/autodoc_sphinx/*.py" = ["D", "E501", "UP", "A", "B", "COM", "EM", "TRY", "PERF", "RUF", "SIM", "FA100"]
"tests/ext/pytest_fixtures/*.py" = ["D", "E501", "UP", "A", "B", "COM", "EM", "TRY", "PERF", "RUF", "SIM", "FA100"]
"tests/ext/docutils/*.py" = ["D", "E501", "UP", "A", "B", "COM", "EM", "TRY", "PERF", "RUF", "SIM", "FA100"]

[tool.pytest.ini_options]
addopts = "--tb=short --no-header --showlocals --doctest-modules --ignore=packages/sphinx-argparse-neo --ignore=packages/sphinx-autodoc-pytest-fixtures --ignore=packages/sphinx-autodoc-docutils"
doctest_optionflags = "ELLIPSIS NORMALIZE_WHITESPACE"
markers = [
  "integration: sphinx integration tests (require full sphinx build)",
]
testpaths = [
  "tests",
  "docs",
  "packages/gp-sphinx/src",
  "packages/sphinx-fonts/src",
  "packages/sphinx-gptheme/src",
]
filterwarnings = [
  "ignore:distutils Version classes are deprecated. Use packaging.version instead.",
  "ignore:The frontend.Option(Parser)? class.*:DeprecationWarning::",
]

[tool.coverage.report]
exclude_also = [
  "def __repr__",
  "raise AssertionError",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
  "if t.TYPE_CHECKING:",
  "class .*\\bProtocol\\):",
  "@(abc\\.)?abstractmethod",
  "from __future__ import annotations",
]
